% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_ABM_Agent.R
\docType{class}
\name{ABM_Agent}
\alias{ABM_Agent}
\title{ABM_Agent Class}
\description{
\code{ABM_Agent} is the core class of \code{netABM} for bundle the data of each agent and its behavior function.
}
\examples{
ABM_Agent$new()
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{.f}}{A function that define the behavior of the agent. Default is \code{NULL}}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ABM_Agent-.save}{\code{ABM_Agent$.save()}}
\item \href{#method-ABM_Agent-print}{\code{ABM_Agent$print()}}
\item \href{#method-ABM_Agent-clone}{\code{ABM_Agent$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ABM_Agent-.save"></a>}}
\if{latex}{\out{\hypertarget{method-ABM_Agent-.save}{}}}
\subsection{Method \code{.save()}}{
Return the the current attributes of the agent
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ABM_Agent$.save()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ABM_Agent-print"></a>}}
\if{latex}{\out{\hypertarget{method-ABM_Agent-print}{}}}
\subsection{Method \code{print()}}{
print the list of field names and the status of the dataset
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ABM_Agent$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ABM_Agent-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ABM_Agent-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ABM_Agent$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
