% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/actAgent_addEdges_random.R
\name{actAgent_addEdges_random}
\alias{actAgent_addEdges_random}
\title{Agent Adds Edges by Randomly Selecting Alters}
\usage{
actAgent_addEdges_random(
  D,
  num_edges = 1,
  alter_candidate = "new",
  which_net = 1,
  value = 1,
  .valueFunction = NULL
)
}
\arguments{
\item{D}{\code{netABM_network} class object}

\item{num_edges}{integer, number of edges to add (default is 1).}

\item{alter_candidate}{character, how to determine the candidates for adding edges (possibilities: \code{"new"} (default), \code{"known"}, or \code{"all"}).
Alternatively, the ID of the candidates can be directly supplied.}

\item{which_net}{integer (index of the network) or character (the name of network) to which edges will be added (default is 1)}

\item{value}{numeric, value of the new edges (default is 1).}

\item{.valueFunction}{function: function to generate edge values (optional. Default is \code{Null}).}
}
\value{
The network of the active agent is updated in place through the \code{R6}'s reference system.
}
\description{
\code{actAgent_addEdges_random} is an agent's behavior function with which user can make agents randomly select alters and add  edges with a defined weight.
}
\details{
\code{actAgent_network} is agent-behavior function which can be used either directly set in \code{.act} argument or
as a part of user-defined agent-behavior of \code{setAgent} family function.

As for the \code{alter_candidate}  argument, following options are available:
\itemize{
\item "new": The candidate of new partner will be selected from those with whom the agent does not have edges.
\item "known": The new partner candidate will be selected from theose with whom the agent already had edges.
\item "all": Any other alters, regardless whether the agent has already has edges or not with them.
\item directly supplying the candidate IDs as character. For multiple candidates, use \code{c( )}.
}

\code{which_net} can be specified either with the integer (i.e., 1, 2, 3...) or the name of the network (e.g., "net1").
If the user has only one network in \code{D} object, the user does not have to care about this argument.

With \code{.valueFunction} argument, the user can supply a function with which the edge value will be generated.
Be ensure that the supplied function only generate one value and does not generate any other values,
because \code{.valueFunction} will be repeated for each selected new agent. If a function generating a random value is supplied,
the resulted edge weight will be different for each new alter. When \code{.valueFunction} is supplied,
the \code{value} argument is ignored. See the Examples below to get the idea about how to supply \code{.valueFunction}.
}
\examples{
# Example 1: Make agents select their partner randomly (leave all arguments as default)
D <- setABM_network(n = 5, .act = list(actAgent_addEdges_random))
D$agents$ID1                    # No edges exists
D$agents$ID1$.act(D)             # Agent adds one edges
D$agents$ID1                    # One edges added

# Example 2: Set edge weight from the normal distribution supplied directly.
 D <- setABM_network(
    n = 5,
    .act = list(actAgent_addEdges_random(.valueFunction = rnorm(n = 1, mean = 0, sd = 1)))
    )

D$agents$ID1                    # No edges exists
D$agents$ID1$.act(D)             # Agent adds one edges
D$agents$ID1                    # One edges added

# Example 3: Set edge weight from the normal distribution supplied via substituted object.
random2 <- substitute(actAgent_addEdges_random(.valueFunction = rnorm(n = 1, mean = 0, sd = 1)))
D <- setABM_network(
  n = 5,
  .act = list(random2))
D$agents$ID1                    # No edges exists
D$agents$ID1$.act(D)             # Agent adds one edges
D$agents$ID1$e                    # One edges added
}
\seealso{
Other actAgent: 
\code{\link{f_nothing}()}
}
\author{
Keiichi Satoh
}
\concept{actAgent}
