% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agrABM.R, R/agrABM_240815.R
\name{agrABM}
\alias{agrABM}
\title{Aggregate and Transform the ABM_D Object for Analysis}
\usage{
agrABM(
  D,
  from_log = FALSE,
  by_field = FALSE,
  attr_long = FALSE,
  incl_attr_ca = FALSE,
  incl_attr_euc = FALSE,
  incl_attr_notes = FALSE,
  ca_arr.ind = FALSE
)

agrABM(
  D,
  from_log = FALSE,
  by_field = FALSE,
  attr_long = FALSE,
  incl_attr_ca = FALSE,
  incl_attr_euc = FALSE,
  incl_attr_notes = FALSE,
  ca_arr.ind = FALSE
)
}
\arguments{
\item{D}{A \code{ABM_D} class object to be processed.}

\item{from_log}{A logical value indicating whether to extract data
from the log \code{TRUE} or directly from the current state \code{FALSE}.
Default is \code{FALSE}.}

\item{attr_long}{A logical value. When \code{TRUE}, agent attributes in each time point is converted into a long format. Default is \code{FALSE}.}

\item{incl_attr_ca}{A logical value. When \code{TRUE}, agent's position in each \code{ca} type field is included in the agent attribute data. Default is \code{FALSE}.}

\item{incl_attr_euc}{A logical value. When \code{TRUE}, agent's position in each \code{euc} type field is included in the agent attribute data. Default is \code{FALSE}.}

\item{incl_attr_notes}{A logical value. When \code{TRUE}, the description in \code{notes}  is included in the agent attribute data. Default is \code{FALSE}.}

\item{ca_arr.ind}{A logical value. When \code{TRUE}, agent's position in each \code{ca} type field is included in a x-y coordinates format. Default is \code{FALSE}.}
}
\value{
A list containing the extracted agent attributes and fields.

A list containing the extracted agent attributes and fields.
}
\description{
The \code{agrABM} function processes a given \code{ABM_D} object
to extract and organize its agent attributes and other fields.
It provides different methods of extracting these components based on the specified parameters.

The \code{agrABM} function processes a given \code{ABM_D} object
to extract and organize its agent attributes and other fields.
It provides different methods of extracting these components based on the specified parameters.
}
\details{
The function processes the data object \code{D} and
extracts its attributes and other fields.
The behavior of the function changes based on the values of
\code{from_log} and \code{by_field} parameters:
\itemize{
\item When \code{from_log} is \code{FALSE}, the function extracts the current agent attributes and the fields directly from \code{D}.
\item When \code{from_log} is \code{TRUE}:
\itemize{
\item If \code{by_field} is \code{FALSE}, the function extracts data for each time point in the log, organizing the output by time.
\item If \code{by_field} is \code{TRUE}, the function organizes the data by fields, optionally using a long format for attributes if \code{attr_long} is \code{TRUE}.
}
}

There are other functions that extract and organize the \code{D} data for some specific purpose (see: "see also" section).

The function processes the data object \code{D} and
extracts its attributes and other fields.
The behavior of the function changes based on the values of
\code{from_log} and \code{by_field} parameters:
\itemize{
\item When \code{from_log} is \code{FALSE}, the function extracts the current agent attributes and the fields directly from \code{D}.
\item When \code{from_log} is \code{TRUE}:
\itemize{
\item If \code{by_field} is \code{FALSE}, the function extracts data for each time point in the log, organizing the output by time.
\item If \code{by_field} is \code{TRUE}, the function organizes the data by fields, optionally using a long format for attributes if \code{attr_long} is \code{TRUE}.
}
}

There are other functions that extract and organize the \code{D} data for some specific purpose (see: "see also" section).
}
\examples{
\dontrun{
# Example usage of agrABM function
result <- agrABM(D = some_D_object)
}
\dontrun{
# Example usage of agrABM function
result <- agrABM(D = some_D_object)
}
}
\seealso{
\code{\link[=get_attr_from_ca]{get_attr_from_ca()}}

\code{\link[=get_attr_from_ca]{get_attr_from_ca()}}
}
