% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agrABM_ca.R
\name{agrABM_ca}
\alias{agrABM_ca}
\title{agrD <- agrABM_ca(D)
agrD <- agrABM_ca(D, from_log = TRUE, attr_long = FALSE)
agrD <- agrABM_ca(D, from_log = TRUE, attr_long = TRUE)}
\usage{
agrABM_ca(D, from_log = FALSE, by_element = FALSE, attr_long = FALSE)
}
\description{
agrD <- agrABM_ca(D)
agrD <- agrABM_ca(D, from_log = TRUE, attr_long = FALSE)
agrD <- agrABM_ca(D, from_log = TRUE, attr_long = TRUE)
}
