% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agrABM_spatNetwork.R
\name{agrABM_spatNetwork}
\alias{agrABM_spatNetwork}
\title{Aggregate Data From A netABM_spatNetwork Object.}
\usage{
agrABM_spatNetwork(D, agr_log = F, attr_long = F)
}
\arguments{
\item{D}{An object of class \code{netABM_spatNetwork}, representing the network ABM simulation data.}

\item{agr_log}{A logical value indicating whether to aggregate the log data. Default is \code{FALSE}.}

\item{attr_long}{A logical value indicating whether to format node attributes in long format. Default is \code{FALSE}.}
}
\value{
A list containing aggregated data from the network ABM simulation,
including node attributes and aggregated networks.
}
\description{
\code{agrABM_spatNetwork} create node attribute data.frame and network matrices,
which are the normal dataset format for the data analysis. The log file can be also converted
either as a separate file for each time slice or aggregated so-called long-format of the panel dataset.
}
\details{
This function aggregates data from a network ABM simulation stored in the 'netABM_network' class object.
If \code{agr_log} is \code{FALSE}, it aggregates the most recent data from
the simulation. If \code{agr_log} is \code{TRUE}, it aggregates data from
the simulation logs. If the log data is not available, it aggregates the most recent data instead.
The \code{attr_long} argument specifies whether to format node attributes in long format.
When \code{TRUE}, it rearranges node attributes into long format, suitable for further analysis.
}
\examples{
# Preparing the data
agent_attr <- data.frame(
   age = c(0, 1, 2, 3, 4),
   sex = c("m","m","m","f","f"))
net <- matrix(1, 5, 5)
move_to_P1 <- function(D){
  D$stage$location$loc[self$ID, self$loc] <- 0
  D$stage$location$loc[self$ID, 1] <- 1
  }
D <- setABM_spatNetwork(agent_n = 5,
                        agent_attr = agent_attr,
                        agent_f = move_to_P1,
                        place_n = 10)
D <- runABM_spatNetwork(D = D, save_log = T)

# Example 1: Simple aggregation of the current dataset
D_agr <- agrABM_spatNetwork(D = D)

# Example 2: Aggregation of the log
D_agr <- agrABM_spatNetwork(D = D, agr_log = TRUE)
D_agr

# Example 3: Aggregation of the log with long-format node attributes
D_agr <- agrABM_spatNetwork(D = D, agr_log = TRUE, attr_long = TRUE)
D_agr$location_network$loc$t2
}
