% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ca_2d_to_3d.R
\name{ca_2d_to_3d}
\alias{ca_2d_to_3d}
\title{Convert a 2D Cellular Automaton (CA) to 3D}
\usage{
ca_2d_to_3d(CA, height)
}
\arguments{
\item{CA}{A 2D matrix representing the cellular automaton.}

\item{height}{An integer specifying the number of layers to replicate the 2D grid along the third dimension.}
}
\value{
A 3D array where the 2D CA grid is replicated along the third dimension.
}
\description{
\code{ca_2d_to_3d} converts a 2D cellular automaton (CA) to
a 3D array by replicating the 2D grid along a third dimension.
}
\details{
The function checks if the input CA is a 2D matrix. If it is, the function replicates this 2D matrix along the third dimension
to create a 3D array. The resulting 3D array has dimensions equal to the original 2D matrix dimensions plus the specified height.
}
\examples{
# Create a 2D CA grid
CA_2d <- matrix(1:9, nrow = 3, ncol = 3)
# Convert the 2D CA grid to a 3D array with height 4
CA_3d <- ca_2d_to_3d(CA_2d, height = 4)

}
\seealso{
Other ca_tools: 
\code{\link{ca_loc}()},
\code{\link{ca_move}()},
\code{\link{get_ca}()}

Other f_tools: 
\code{\link{ca_loc}()},
\code{\link{ca_move}()},
\code{\link{calc_logistic}()},
\code{\link{calc_softmax}()},
\code{\link{f_nothing}()},
\code{\link{get_ca}()}
}
\author{
Keiichi Satoh
}
\concept{ca_tools}
\concept{f_tools}
