% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ca_loc.R
\name{ca_loc}
\alias{ca_loc}
\title{Get Location of an Agent in a Cellular Automaton (CA)}
\usage{
ca_loc(which_ca = NULL, arr.ind = TRUE, ID = NULL, D = NULL)
}
\arguments{
\item{which_ca}{A name specifying which CA to use from the \code{D} list.
If \code{NULL}, the first ca field found in the \code{D} is automatically found.}

\item{arr.ind}{A logical value indicating whether to return the array indices
(row and column) of the agent's location. If FALSE, returns the position as
an integer-valued index. Default is \code{TRUE}.}

\item{ID}{The ID of the agent whose location is to be found.
If this is remain to be the default \code{NULL},
the current agent's ID is automatically retrieved.}

\item{CA}{An optional matrix or array representing the CA.
If provided, this CA is used directly. Default is \code{NULL}.}
}
\value{
The function returns the location of the specified agent ID in the CA.
The format of the return value depends on the \code{arr.ind} parameter:
\itemize{
\item If \code{arr.ind = TRUE}, returns the row and column indices of the agent's location.
\item If \code{arr.ind = FALSE}, returns the position as an integer-valued index.
}
}
\description{
This function retrieves the location of a specified agent ID in a cellular automaton (CA).
}
\details{
The function first checks if a CA is directly provided through
the \code{CA} parameter. If not, it retrieves the CA from the \code{D} list
using the specified \code{which_ca} index.
It then finds and returns the location of the specified agent ID in the CA.
}
\examples{

# Example 1: use get_ca_loc as an action of an agent
f <- function(){
  get_ca_loc()
}
D <- setABM(agent_n = 3, ca = 1, agent_f = f)
D$agent$ID1$.f(D = D)

# Example 2: retireve the agent location directly
get_ca_loc(CA = D$ca, ID = 1)
}
\seealso{
Other ca_tools: 
\code{\link{ca_2d_to_3d}()},
\code{\link{ca_move}()},
\code{\link{get_ca}()}

Other f_tools: 
\code{\link{ca_2d_to_3d}()},
\code{\link{ca_move}()},
\code{\link{calc_logistic}()},
\code{\link{calc_softmax}()},
\code{\link{f_nothing}()},
\code{\link{get_ca}()}
}
\author{
Keiichi Satoh
}
\concept{ca_tools}
\concept{f_tools}
