% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_logistic.R
\name{calc_logistic}
\alias{calc_logistic}
\alias{choose_logistic}
\title{Calculate Probabilities Using Logistic Function}
\usage{
calc_logistic(
  formula,
  data = NULL,
  standardize = FALSE,
  base_prob = NULL,
  randomness = rnorm(sd = 0.01)
)

choose_logistic(
  formula,
  data = NULL,
  standardize = FALSE,
  base_prob = NULL,
  randomness = rnorm(sd = 0.01)
)
}
\arguments{
\item{formula}{A formula specifying the model.
The left-hand side (LHS) can optionally specify the labels,
and the right-hand side (RHS) specifies the predictors.}

\item{data}{A data frame containing the variables in the formula.}

\item{standardize}{A logical value indicating whether to standardize numeric predictors.
Default is \code{FALSE}.}

\item{base_prob}{A constant probability that will be added as a intercept into the formula.
The default \code{NULL} adds no intercept into the formula.}

\item{randomness}{A funtion that generates random number.
The default \code{rnorm(sd = 0.01)} adds around 1\% randomness into
the probability in addition to specified formula.}
}
\value{
\code{calc_logistic} returns a probability.
\code{choose_softmax} returns TRUE or FALSE based on the logistic probability.
}
\description{
This function calculates the probabilities of choosing the dependent variable
based on the logistic function for a given formula and dataset.
}
\details{
\code{calc_logistic}  evaluates the specified formula in the context of the
provided data.
It computes the probabilities for the predictors on the RHS of the formula.
If standardize is \code{TRUE}, numeric predictors are standardized
before calculating the probabilities.

\code{choose_logistic} samples from the choices based on these probabilities.
}
\examples{
# Create sample data
n <- 10
dat <- data.frame(
   Y = LETTERS[1:10],
   X1 = rnorm(n),
   X2 = rnorm(n)
 )

# The left-hand formula is used to index the result
calc_softmax(Y ~ X1 * X2, data = dat, standardize = FALSE)
# Without the left-hand specification, the result is returned with the row index
calc_softmax(~ X1 * X2, data = dat, standardize = FALSE)

# Choose based on softmax probabilities
choose_softmax(Y ~ X1 * X2, data = dat, standardize = FALSE, n = 2)
dat <- data.frame(
Y = LETTERS[1:3],
X1 = rnorm(n = 3),
X2 = rnorm(n = 3)
)

calc_logistic(Y ~ X1 + X2, data = dat, base_prob = 0.1)
calc_logistic(~ X1 + X2, data = dat, base_prob = 0.1)
calc_logistic(~ 0, data = dat, base_prob = 0.05)
calc_logistic(~ 0, data = dat, base_prob = 0.05, randomness = rnorm(sd = 0.01))
choose_logistic(~ 0, data = dat, base_prob = 0.05, randomness = rnorm(sd = 0.01))
}
\seealso{
Other f_tools: 
\code{\link{ca_2d_to_3d}()},
\code{\link{ca_loc}()},
\code{\link{ca_move}()},
\code{\link{calc_softmax}()},
\code{\link{f_nothing}()},
\code{\link{get_ca}()}
}
\author{
Keiichi Satoh
}
\concept{f_tools}
