% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/D_to_statnet.R
\name{D_to_statnet}
\alias{D_to_statnet}
\title{Convert netABM Object to statnet}
\usage{
D_to_statnet(D, log = F, duplicated_edges = "sum", which_net = 1)
}
\arguments{
\item{D}{A \code{netABM} object.}

\item{log}{Logical, indicating whether to convert the log of \code{netABM} object into a list of \code{statnet} network objects. Default is \code{FALSE}.}

\item{duplicated_edges}{Character string indicating how duplicated edges should be handled. Options are \code{"sum"}, \code{"min"}, \code{"max"}, or \code{"mean"}. Default is \code{"sum"}.}

\item{which_net}{Integer or character string specifying which network to convert if there are multiple networks in the ABM object. Default is \code{1}.}
}
\value{
A statnet network object if \code{log} is \code{FALSE}, or a list of statnet network objects if \code{log} is \code{TRUE}.
}
\description{
\code{D_to_statnet} converts \code{netABM} object to a statnet network object.
}
\details{
\code{statnet}(see: https://statnet.org/)  is a popular suite package for statistical social network analysis which takes \code{network} class object as an input.
The \code{D_to_statnet} function internally calls \code{agrABM_network} to create an aggregated dataset of D and then passes
it to \code{statnet}'s \code{network} function to create an \code{network} class object.
}
\examples{
# preparing the dataset
D <- setABM_network(n = 5,
.act = actAgent_addEdges_random)
D <- runABM_network(D,
                    .stopCondition = stopABM_times(simTimes = 5),
                    save_log = TRUE)

# Example 1: Convert the current D object into statnet's network object
D_net <- D_to_statnet(D)

# Example 2: Convert the log of D object into the list of statnet's network object
D_net <- D_to_statnet(D, log = TRUE)
}
\seealso{
\code{\link[=agrABM_network]{agrABM_network()}}, \code{\link[network:network]{network::network()}}
}
