% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_attr.R
\name{get_attr}
\alias{get_attr}
\title{Get A Agents' Attributes From A netABM Object.}
\usage{
get_attr(D, which_attr = NULL)
}
\arguments{
\item{D}{An object of class \code{netABM}.}

\item{which_attr}{An integer or character string indicating the agents' attribute(s).
The default is \code{NULL}, resulting in selecting the whole attributes.}
}
\value{
A vector if only one attribute is queried and a data.frame if multiple attributes.
}
\description{
\code{get_attr} creates a vector and data.frame of agents' attributes.
}
\details{
Upon writing the function of the agent's action, there is often the case
that the action depends on the agents' attribute(s). For example,
if the agent's probability to form an edge with an alter depends on the alters' attributes,
the user firstly need to get the vector or data.frame of the agents' attributes as an input.
This function can be used for such purpose among others, but can be used
for other purposes.
}
\examples{
# Create the netABM_network object
node_attr <- data.frame(
  age = 1:5,
  sex = c("m","m","m","f","f"))
D <- setABM_network(n = 5,
                    node_attr = node_attr)

get_attr(D)

# example 2: get a selected attribute
get_attr(D, "sex")

# example 3: get a selected attribute by its position
get_attr(D, 2)

# example 4: get multiple selected attributes
get_attr(D, c("age","sex"))
}
\seealso{
Other act_tools: 
\code{\link{add_edges_self}()},
\code{\link{get_net}()}
}
\concept{act_tools}
