% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_attr_from_ca.R
\name{get_attr_from_ca}
\alias{get_attr_from_ca}
\title{Retrieve Attributes from CA}
\usage{
get_attr_from_ca(
  D,
  ID_ca,
  attr_ca = NULL,
  with_agent_attr = FALSE,
  incl_posit = FALSE,
  from_log = FALSE,
  attr_long = FALSE,
  which_attr = NULL
)
}
\arguments{
\item{D}{The \code{ABM_D} object containing \code{ca} type field.}

\item{ID_ca}{The identifier for the \code{ca} type field within the \code{D}.}

\item{attr_ca}{A vector of attribute names to retrieve.
If \code{NULL}, all \code{ca} type fields within \code{D} except for \code{ID_ca}
will be used.}

\item{with_agent_attr}{A logical value indicating whether to include agent attributes in the output.}

\item{incl_posit}{A logical value indicating whether to include the position information of the \code{ca}.}

\item{from_log}{A logical value indicating whether to retrieve data from the log.}

\item{attr_long}{A logical value indicating whether to output the attributes in long format.}

\item{which_attr}{Charactor vector of attributes. The default \code{NULL} retrieves all agent's attributes.}
}
\value{
A data frame containing the retrieved CA attributes.
If \code{from_log} is \code{TRUE}, a list of data frames
is returned, one for each time point in the log.
If \code{attr_long} is \code{TRUE}, a single data frame in long format
is returned.
}
\description{
This function retrieves specific attributes from a \code{ca} (cellular automaton) type field
within the given \code{D} object.
The attributes can be retrieved from the current data or from log data,
depending on the parameters provided.

In some cases, the user firstly defines the attribute of the {ca} by using the \code{ca} type field
and wants to extract this \code{ca} field attribute of the cell in which agents locates.
This function can be used for such scenario.
}
\examples{
# define ca attributes
zone <- matrix(c(1,1,
                 2,2), 2, 2, byrow = TRUE)
D <- setABM(agent_n = 3, ca = list(ca = 2, zone = zone))
# View where each agent locates
D$ca
# Extract the zone number where each agent locates
get_attr_from_ca(D = D, ID_ca = "ca", attr_ca = "zone")
}
\seealso{
\code{\link[=agrABM]{agrABM()}}
}
