% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ca_by_attr.R
\name{get_ca_by_attr}
\alias{get_ca_by_attr}
\title{Get Cellular Automata by A Specified Attribute}
\usage{
get_ca_by_attr(
  D = NULL,
  which_ca = 1,
  which_attr = 1,
  from_log = FALSE,
  CA = NULL,
  attr_list = NULL,
  exclude_value = 0
)
}
\arguments{
\item{D}{A \code{netABM} class object containing CA and attributes. This parameter is required if \code{CA} and \code{attr_list} are not provided directly.}

\item{which_ca}{Either an index or name specifying which CA to extract from \code{D}. Default is \code{1}.}

\item{which_attr}{Either an index or name specifying which attribute to use for extraction. Default is \code{1}.}

\item{from_log}{A logical value indicating whether to extract data from the log (\code{TRUE}) or from the current state (\code{FALSE}). Default is \code{FALSE}.}

\item{CA}{A list or array of CA data. If provided, this CA data is used directly, and \code{D} is not required.}

\item{attr_list}{A list of attribute vectors. If provided, this attribute list is used directly, and \code{D} is not required.}

\item{exclude_value}{A value in the CA data that should be excluded and replaced with \code{NA}. Default is \code{0}.}
}
\value{
A list or array of CA data with the specified attributes applied. If only one CA data set is retrieved, it is returned as an array.
}
\description{
The \code{get_ca_by_attr} function retrieves cellular automata (CA) based on specified attributes
from a data object \code{D} or directly from provided CA and attribute list.
The function can also extract CA from a \code{log} if specified.
}
\details{
The function processes the data object \code{D} or the provided \code{CA} and \code{attr_list}
to retrieve the CA data based on the specified attributes. The function allows for flexibility in how the CA data and attributes are specified,
either directly or via the data object \code{D}.
}
\examples{
\dontrun{
# Example usage of get_ca_by_attr function with a data object
result <- get_ca_by_attr(D = some_data_object)

# Example usage with data from log
result <- get_ca_by_attr(D = some_data_object, from_log = TRUE)

# Example usage with direct CA and attribute list
result <- get_ca_by_attr(CA = some_data_object$ca$room, D = some_data_object)
result <- get_ca_by_attr(CA = some_data_object$ca$room, attr_list = list(c("F", "F", "F", "J", "J")))
}

}
\seealso{
\code{\link{agrABM_ca}} for a related function to process and extract attributes and CA components.
}
