% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ca_loc.R
\name{get_ca_loc}
\alias{get_ca_loc}
\title{Get Location of an Agent in a Cellular Automaton (CA)}
\usage{
get_ca_loc(D = NULL, ID = NULL, which_ca = 1, CA = NULL, arr.ind = TRUE)
}
\arguments{
\item{D}{A \code{netABM} class object set by \code{setABM_ca} function. Default is \code{NULL}.}

\item{ID}{The ID of the agent whose location is to be found.}

\item{which_ca}{An index or the name specifying which CA to use from the \code{D} list.
Default is \code{1}.}

\item{CA}{An optional matrix or array representing the CA. If provided, this CA is used directly.
Default is \code{NULL}.}

\item{arr.ind}{A logical value indicating whether to return the array indices
(row and column) of the agent's location. If FALSE, returns the position as
an integer-valued index. Default is \code{TRUE}.}
}
\value{
The function returns the location of the specified agent ID in the CA. The format of the return value depends on the \code{arr.ind} parameter:
\itemize{
\item If \code{arr.ind = TRUE}, returns the row and column indices of the agent's location.
\item If \code{arr.ind = FALSE}, returns the position as an integer-valued index.
}
}
\description{
This function retrieves the location of a specified agent ID in a cellular automaton (CA).
The function can be memoized to cache results for faster repeated access.
}
\details{
The function first checks if a CA is directly provided through the \code{CA} parameter. If not, it retrieves the CA from the \code{D} list using the specified \code{which_ca} index.
It then finds and returns the location of the specified agent ID in the CA.
}
\examples{
# Create a sample CA data
ca <- init_ca(dim = c(3, 3), agent_n = 5)
D <- setABM_ca(agent_n = 5, ca = ca)

# Get the location of ID = 1
get_ca_loc(D = D, ID = 1, which_ca = "ca")                   # Get the row and column
get_ca_loc(D = D, ID = 1, which_ca = "ca", arr.ind = FALSE)  # Get the position as integer-valued index

# Supply CA directly
get_ca_loc(CA = ca, ID = 1)

}
\seealso{
Other ca_tools: 
\code{\link{ca_2d_to_3d}()},
\code{\link{ca_move}()}

Other f_tools: 
\code{\link{ca_2d_to_3d}()},
\code{\link{ca_move}()},
\code{\link{calc_softmax}()},
\code{\link{get_attr}()}
}
\author{
Keiichi Satoh
}
\concept{ca_tools}
\concept{f_tools}
