% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_net.R
\name{get_net}
\alias{get_net}
\title{Get A Whole Network From A netABM Object.}
\usage{
get_net(D, which_net = 1, space = FALSE)
}
\arguments{
\item{D}{An object of class \code{netABM}.}

\item{which_net}{An integer or character string indicating the network to convert.}

\item{space}{logical, whether network to get locates at space level (\code{TRUE}).
The default is \code{FALSE} and the agent level network is returned.}
}
\value{
A whole network matrix created from \code{D}
}
\description{
\code{get_net} creates a whole network matrix from the input \code{netABM} class D object.
}
\details{
Upon writing the function of the agent's action, there is often the case
that the action depends on the condition of the whole network. For example,
if the agent's probability to form an edge with an alter depends on the indegree of the
alter, the user firstly calculate the indegree of the alters which typically
needs an whole network as an input.
This function can be used for such purpose among others, but can be used
for other purposes.
}
\examples{
# Make a sample dataset
D <- setABM_network(n = 5,
                    networks = list(net1 = sna::rgraph(5, tprob = 0.2),
                                    net2 = sna::rgraph(5, tprob = 0.5)))
# Example 1: get the whole network(default)
get_net(D)

# Example 2: get the 2nd whole network
get_net(D, 2)

# Example 3: get the 2nd whole network by its name
get_net(D, "net2")

}
\seealso{
Other act_tools: 
\code{\link{add_edges_self}()},
\code{\link{get_attr}()}
}
\concept{act_tools}
