% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_graph.R
\name{grid_graph}
\alias{grid_graph}
\title{Generate a grid graph}
\usage{
grid_graph(n, k)
}
\arguments{
\item{n}{The number of vertices}

\item{k}{The number of neighboring vertices connected to each vertex. Note that the resulting neighboring vertices will be 2*k.}
}
\value{
The grid graph represented in adjacency matrix format
}
\description{
This function generates a grid graph with the given number of vertices and degree.

This function generates a special case of grid graph where each vertices are corrected
to their given number of neighboring vertices and there is no "end" vertices because the
the vertices of the last ID again connects to the vertices with the first ID.
This graph can be useful to set a initial condition of the spatial networks where
agent on the given vertices can move to the neighboring vertices.
This graph can be seen as a generalization of the cellular automaton
to the given dimension. Note that each vertices has 2*k neighbors.
Concretely, by setting the following number of k,
the movable direction of the agents on a vertices changes from 2 (k = 1), 4 (k = 2),
6 (k = 3), and 8 (k = 4).
}
\examples{
# k = 1:cicle
grid_graph(10, 1)

# k = 2: A square Generate a 4x4 grid graph
grid_graph(10, 2)
}
