% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_agent.R
\name{init_agent}
\alias{init_agent}
\title{Create a set of ABM_Agent class agents}
\usage{
init_agent(
  n = NULL,
  attr = NULL,
  act_FUN = NULL,
  active_binding_field = NULL,
  ID_start = 1
)
}
\value{
A 3D array where the 2D CA grid is replicated along the third dimension.
}
\description{
\code{init_agent} create ABM_Agent class agents, which can be
an input to setABM.
}
\details{
The function checks if the input CA is a 2D matrix. If it is, the function replicates this 2D matrix along the third dimension
to create a 3D array. The resulting 3D array has dimensions equal to the original 2D matrix dimensions plus the specified height.
}
\author{
Keiichi Satoh
}
\concept{init}
