% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lapply_mms.R
\name{lapply_mms}
\alias{lapply_mms}
\title{memoised lapply}
\usage{
lapply_mms(X, FUN, ...)
}
\value{
a list.
}
\description{
\code{memoised lapply} is the memoised version of lapply. Any other arguments remains the same as the original \code{lapply}.
}
\details{
Technically, it uses a memoised function; therefore,
as long as the contents of the input object remains the same, the return is remembered for 15 minutes and then automatically
deleted from the user's temp file.
}
\seealso{
\code{\link[=lapply]{lapply()}} for the arguments.
}
\author{
Keiichi Satoh
}
\concept{netABM_misc}
