% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapplyly.R
\name{mapplyly}
\alias{mapplyly}
\title{library(rlang)}
\usage{
mapplyly(FUN, which_list = 1, .combine = NULL)
}
\description{
library(rlang)
}
\details{
mapplyly(sum(mat, na.rm = FALSE))
}
\section{INPUT}{
mat <- list(matrix(1:9,3,3),
matrix(11:19, 3,3),
matrix(21:29, 3,3))
}

