% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netABM_network.R, R/print.netABM_network.R
\name{netABM_network}
\alias{netABM_network}
\alias{print.netABM_network}
\title{netABM_network}
\usage{
netABM_network(x, ...)

netABM_network(x, ...)

print.netABM_network(D)
}
\arguments{
\item{D}{A netABM_netork object}
}
\description{
netABM_network

netABM_network framework
}
\examples{
mat <- sna::rgraph(5)
D <- setABM_network(agent_n = 5, agent_network = mat)
print.netABM_network(D)

print(D)    # do the same
}
