% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/para_run.R
\name{para_run}
\alias{para_run}
\title{library(doParallel)
library(doRNG)
library(foreach)
library(rlang)}
\usage{
para_run(FUN, cores = 2, n_sim = 3, seed = NULL, .export = NULL)
}
\description{
out <- para_run(FUN = setABM_ca(agent_n = 6), cores = 3, n_sim = 3, seed = 1)
}
