% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ca.R
\name{plot_ca}
\alias{plot_ca}
\title{library(reshape)
library(ggplot2)
library(RColorBrewer)}
\usage{
plot_ca(D = NULL, which_ca = 1, CA = NULL, attr = NULL, show.ID = TRUE)
}
\description{
\subsection{Example}{

attr <- data.frame(X1 = as.character(c(1,1,1,2,2)))
D <- setABM_ca(agent_n = 5, agent_attr = attr)
plot_ca(D, attr = "X1")
}
}
