% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.netABM_ca.R
\name{print.netABM_ca}
\alias{print.netABM_ca}
\title{Printing the netABM_ca class}
\usage{
\method{print}{netABM_ca}(D)
}
\arguments{
\item{D}{A netABM_ca object}
}
\description{
Printing the netABM_ca class
}
\examples{
D <- setABM_ca(agent_n = 5)
print.netABM_ca(D)

print(D)    # do the same
}
