% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.netABM_network.R
\name{print.netABM_network}
\alias{print.netABM_network}
\title{Printing the netABM_network class}
\usage{
print.netABM_network(D)
}
\arguments{
\item{D}{A netABM_netork object}
}
\description{
Printing the netABM_network class
}
\examples{
mat <- sna::rgraph(5)
D <- setABM_network(agent_n = 5, agent_network = mat)
print.netABM_network(D)

print(D)    # do the same
}
