% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.netABM_spatNetwork.R
\name{print.netABM_spatNetwork}
\alias{print.netABM_spatNetwork}
\title{Printing the netABM_spatNetwork class}
\usage{
\method{print}{netABM_spatNetwork}(D)
}
\arguments{
\item{D}{A netABM_spatNetwork object}
}
\description{
Printing the netABM_spatNetwork class
}
\examples{

D <- setABM_spatNetwork(agent_n = 5, place_n = 10)
print.netABM_spatNetwork(D)
print(D)

D    # do the same
}
