% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_grouping.R
\name{random_grouping}
\alias{random_grouping}
\title{Randomly group IDs into specified group sizes}
\usage{
random_grouping(id, size = 2, remainder_meth = c("omit", "keep", "isolate"))
}
\arguments{
\item{id}{A vector of IDs to be grouped.}

\item{size}{An integer specifying the size of each group. Defaults to 2.}

\item{remainder_meth}{A character string indicating how to handle IDs that do not fit evenly into groups.
Options are:
\itemize{
\item \code{"omit"}: Discard the remainder IDs.
\item \code{"keep"}: Add the remainder IDs as a single group, padded with \code{NA} if necessary.
\item \code{"isolate"}: Place each remainder ID in its own group, padded with \code{NA} to match the group size.
}
Defaults to \code{"omit"}.}
}
\value{
A matrix where each row represents a group, and columns contain IDs or \code{NA} if the group is smaller than the specified size.
}
\description{
This function takes a vector of IDs, shuffles them randomly, and divides them into groups of a specified size.
If the total number of IDs is not evenly divisible by the group size, the behavior for handling the remainder
can be controlled via the \code{remainder_meth} argument.
}
\examples{
random_grouping(id = 1:8, size = 2)
random_grouping(id = 1:8, size = 3, remainder_meth = "omit")
random_grouping(id = 1:8, size = 3, remainder_meth = "keep")
random_grouping(id = 1:8, size = 3, remainder_meth = "isolate")

}
