% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectAgent_all.R
\name{selectAgent_all}
\alias{selectAgent_all}
\title{Selecting All Agents in netABM objects}
\usage{
selectAgent_all(D)
}
\arguments{
\item{D}{a \code{netABM} object.}
}
\value{
a vector of the IDs of agents in the \code{netABM} class object D.
}
\description{
\code{selectAgent_all} returns all agents in the netABM objects.
}
\details{
A typicall usage of \code{selectAgent_all} is setting this function as an argument of \code{runABM_network},
so that all agents acts simultaneously per time. Technically, it uses a memoised function; therefore,
as long as the original D object remains the same, the return is remembered for 15 minutes and then automatically
deleted from the user's temp file.
}
\examples{
D <- setABM_network(n = 5, .act = list(NULL))
selectAgent_all(D = D)
}
\author{
Keiichi Satoh
}
\concept{selectAgent}
