% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setABM_network.R
\name{setABM_network}
\alias{setABM_network}
\title{Setting netABM_network Objects}
\usage{
setABM_network(n, node_attr = NULL, .act = NULL, networks = NULL)
}
\arguments{
\item{n}{Number of agents}

\item{node_attr}{vector/data.frame/list of attributes of agents (default: \code{NULL})}

\item{.act}{A user-defined or built-in function object representing agent's actions or list of functions for each agent. The default value \code{NULL} will result in setting \code{actAgent_nothing} to all agents.}

\item{networks}{Single n times n matrix or dataframe, or list of multiple matrices. The default value \code{NULL} will result in creating a n times n matrix without any edges.}
}
\value{
a D object (see Details)
}
\description{
\code{setABM_network} constructs a \code{netABM_network} object for running ABM.
}
\details{
\code{setAgent_network} is a constructor of \code{netABM_network} object (D)
which has \code{agents}, \code{time}(set as 1), and \code{log} (set as NA) as a list format.
\code{agents} contains each agent's attribute, networks, and actions as a \code{R6} class object from package \code{R6}.

Each agent automatically get its \code{ID} and \code{act_label} and store them as their attributes.
The latter \code{act_label} is put from the supplied object name of \code{.act}.

There are two ways to set \code{.act}.
The first way is to write the user's own function of agent's actions.
Upon writing an original function, be sure to set \code{D} as the first argument without any default;
otherwise agent's action does not reflect dynamically to the changing \code{D} object during the simulation.
\code{self} is a reserved for indicating the agent themselves.

The second way of setting \code{.act} is to use a built-in function of this package.
If user wants to modify some argument, supply it as a form: \code{function_name(x = a new value)}.

For now, \code{setABM_network} only supports directed networks.
}
\examples{
node_attr <- data.frame(
age = c(0, 1, 2, 3, 4),
sex = c("m","m","m","f","f"))
network <- matrix(1, 5, 5)
agent_get_older <- function(D, b = 1){self$age <- b*self$age + 1}

D <- setABM_network(n = 5,
                   node_attr = node_attr,
                   networks = network,
                   .act = agent_get_older)
}
\author{
Keiichi Satoh
}
\concept{setABM}
