% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stage_by_attr.R
\name{stage_by_attr}
\alias{stage_by_attr}
\title{Replace Agent IDs on a Stage Field with an Agent Attribute}
\usage{
stage_by_attr(stage_name, agents_name, attr_name, G = NULL)
}
\arguments{
\item{stage_name}{A character string specifying the name of the stage
in the \code{ABM_G} object.}

\item{agents_name}{A character string specifying the name of the agents
object in the \code{ABM_G} object.}

\item{attr_name}{A character string specifying the name of the attribute
to replace the agent IDs with.}

\item{G}{An ABM object containing the agents and stage.
If \code{NULL}, it uses \code{self} to refer to the ABM environment. For example,
if users uses this function as an \code{active_binding_field} in \code{G} object,
leave this field as default.}
}
\value{
A matrix where agent IDs in the specified stage have been replaced with the values of the specified attribute.
}
\description{
Replaces the agent IDs on a stage field with a
specified agent attribute. The function modifies the stage field in \code{G} object
by substituting agent IDs with the values of a given attribute from the agents.
}
\examples{
persons <- init_agent(n = 3, attr_df = data.frame(sex = c("f","f","m")))
stage <- matrix(c(1:3, 0), 2, 2)
G <- setABM(agents = persons, stage = stage)
stage_by_attr(stage_name = "stage", agents_name = "persons", attr_name = "sex", G = G)

}
