% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_ABM_D.R
\docType{class}
\name{ABM_D}
\alias{ABM_D}
\title{ABM_D Class}
\description{
\code{ABM_D} is the core class of \code{netABM} that bundle datasets for running ABM.
}
\details{
ABM class
}
\examples{
ABM_D$new()
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{agent}}{list of the agents}

\item{\code{time}}{current time of the dataset}

\item{\code{log}}{list of the past dataset}

\item{\code{notes}}{list of the notes. Default is \code{NULL}}

\item{\code{.f}}{A function that control the fields of ABM_D. Default is \code{NULL}}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ABM_D-.save}{\code{ABM_D$.save()}}
\item \href{#method-ABM_D-print}{\code{ABM_D$print()}}
\item \href{#method-ABM_D-.field_type}{\code{ABM_D$.field_type()}}
\item \href{#method-ABM_D-.agent_attr}{\code{ABM_D$.agent_attr()}}
\item \href{#method-ABM_D-clone}{\code{ABM_D$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ABM_D-.save"></a>}}
\if{latex}{\out{\hypertarget{method-ABM_D-.save}{}}}
\subsection{Method \code{.save()}}{
Save the current status of the fields into log
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ABM_D$.save()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ABM_D-print"></a>}}
\if{latex}{\out{\hypertarget{method-ABM_D-print}{}}}
\subsection{Method \code{print()}}{
print method of the \code{ABM_D} class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ABM_D$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ABM_D-.field_type"></a>}}
\if{latex}{\out{\hypertarget{method-ABM_D-.field_type}{}}}
\subsection{Method \code{.field_type()}}{
get the field type of the \code{ABM_D} class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ABM_D$.field_type()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ABM_D-.agent_attr"></a>}}
\if{latex}{\out{\hypertarget{method-ABM_D-.agent_attr}{}}}
\subsection{Method \code{.agent_attr()}}{
get the attribute of the agents in \code{ABM_D} class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ABM_D$.agent_attr(from_log = FALSE, attr_long = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{from_log}}{logical. If the attribute is retrieved from log.}

\item{\code{attr_long}}{logical. If the attribute from the log is converted into the long format.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ABM_D-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ABM_D-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ABM_D$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
