% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ca_move.R
\name{ca_move}
\alias{ca_move}
\title{D <- setABM_ca(agent_n = 5)
ca_move(D, ID = 1, where_to = 5, which_ca = 1)
D$ca}
\usage{
ca_move(D, ID, where_to, which_ca = 1)
}
\description{
D <- setABM_ca(agent_n = 5)
ca_move(D, ID = 1, where_to = 5, which_ca = 1)
D$ca
}
