% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ca_neib.R
\name{ca_neib}
\alias{ca_neib}
\title{Get Neighboring Cells in a Cellular Automaton}
\usage{
ca_neib(
  which_ca = NULL,
  neib_type = "Moore",
  torus = FALSE,
  which_attr = NULL,
  neib_size = 1,
  D = NULL,
  ID = NULL,
  loc = NULL,
  FUN = NULL,
  ...
)
}
\arguments{
\item{which_ca}{The index or name of the CA grid to use within the \code{D} object.
Defaults is \code{NULL}, resulting in selecting the first "ca" object found by
\code{D$.field_type()} method.}

\item{neib_type}{The type of neighborhood to use,
either \code{"Moore"} or \code{"Neumann"}. Defaults to \code{"Moore"}.}

\item{torus}{Logical, indicating whether the grid should be treated as a torus.
Defaults to \code{FALSE}.}

\item{which_attr}{A character vector specifying the attributes of the
neighboring agents to retrieve. If \code{NULL}, only the IDs of neighboring
cells are returned.}

\item{neib_size}{A numeric vector specifying the distance of neighbors considered
from the agent. If a single vector is provided, it will copied to each dimension.
Default is \code{1}.}

\item{D}{A \code{ABM_D} class object. The default \code{NULL} results in retrieving
D from the parent.frame, which will be usual usage.}

\item{ID}{The ID of the target agent. The default \code{NULL} results in retrieving
the ID from the parent.frame, which will be usual usage.}

\item{loc}{The location of the target cell as a 1*CA's dimension matrix format
(e.g., matrix(c(row_ID, column_ID)), 1, 2))
If not provided, it will be retrieved from the \code{ID}'s location.}

\item{FUN}{An optional function to apply to the attributes of neighboring
agents to calculate the specified score.
If \code{NULL}, the attributes are returned as-is.}

\item{...}{Additional arguments passed to \code{FUN}.}
}
\value{
Depending on the input parameters:
\itemize{
\item If \code{which_attr} is \code{NULL}, a vector of neighboring cell IDs is returned.
\item If \code{which_attr} is provided, a data frame of the specified
attributes of neighboring agents is returned.
\item If both \code{which_attr} and \code{FUN} are provided,
the result of applying \code{FUN} to the attributes is returned.
}
}
\description{
This function retrieves the ID of the neighboring cells in a
cellular automaton (CA) grid, or either just retrieve or calculate their attributes.
It can handle both Moore and Neumann neighborhood types and
can also handle toroidal grids.
}
\examples{
# creating a CA
CA <- matrix(1:16, 4, 4)
attr <- 1:16
# Get attributes of neighboring agents
ca_neib(ID = 1, CA = CA, attr = attr)
# Apply a function to the attributes of neighboring agents
ca_neib(ID = 1, CA = CA, attr = attr, FUN = mean)
D <- setABM(agent_n = 5, agent_attr = data.frame(age = 1:5), ca = 1)
ca_neib(D = D, ID = 1)
ca_neib(D = D, ID = 1, which_attr = "age")
ca_neib(D = D, ID = 1, which_attr = "age", FUN = mean)
}
