% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_softmax.R
\name{calc_softmax}
\alias{calc_softmax}
\title{package
library(rlang)}
\usage{
calc_softmax(formula, data = NULL, standardize = FALSE)
}
\description{
package
library(rlang)
}
\details{
calc_softmax(Y ~ X1 * X2, data = dat, standardize = FALSE)
calc_softmax(  ~ X1 * X2, data = dat, standardize = FALSE)
}
\section{EXAMPLE}{
n <- 10
dat <- data.frame(
Y = LETTERS\link{1:10},
X1 = rnorm(n),
X2 = rnorm(n)
)
}

