% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ca_by_attr.R
\name{get_ca_by_attr}
\alias{get_ca_by_attr}
\title{get_ca_by_attr(D)
get_ca_by_attr(D, from_log = TRUE)
get_ca_by_attr(CA = D$ca$room, D = D)
get_ca_by_attr(CA = D$ca$room, attr_vec_list = list(c("F","F","F","J","J")))}
\usage{
get_ca_by_attr(
  D = NULL,
  which_ca = 1,
  which_attr = 1,
  from_log = FALSE,
  CA = NULL,
  attr_list = NULL,
  exclude_value = 0
)
}
\description{
get_ca_by_attr(D)
get_ca_by_attr(D, from_log = TRUE)
get_ca_by_attr(CA = D$ca$room, D = D)
get_ca_by_attr(CA = D$ca$room, attr_vec_list = list(c("F","F","F","J","J")))
}
