% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netABM_network.R
\name{print.netABM_network}
\alias{print.netABM_network}
\title{print.netABM_network - Print Method for netABM_network}
\usage{
\method{print}{netABM_network}(D)
}
\arguments{
\item{D}{A netABM_network object}
}
\description{
print.netABM_network - Print Method for netABM_network
}
