% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stopABM_times.R
\name{stopABM_times}
\alias{stopABM_times}
\title{Stop ABM Simulation by Runtime}
\usage{
stopABM_times(D, simTimes = 1)
}
\arguments{
\item{D}{The D object. Note that this will be supplied internally within \code{runABM}. Therefore, user needs to leave this argument as it is.}

\item{simTimes}{Number of run (Default: 1)}
}
\value{
TRUE (if the simulation has run for the defined number of times specified by simTimes)
}
\description{
\code{stopABM_times} returns TRUE when the ABM simulation continues after the defined time.
}
\examples{
# This function is supposed to be used in tandem with runABM family function.

# D <- setABM_network(n = 5)
# runABM_network(D=D, .stopCondition = stopABM_simTimes(simTimes = 5))
}
\author{
Keiichi Satoh
}
\concept{stopABM}
